restructured project, added makeStyle
