+++ jsconfig.json
@@ -1,6 +1,6 @@
 {
   "compilerOptions": {
-  "baseUrl": "src"
-},
+    "baseUrl": "src"
+  },
   "include": ["src"]
 }
