+++ src/components/CopyButton.js
@@ -1,10 +1,19 @@
 import React, {useState} from 'react';
 import Button from '@material-ui/core/Button';
-import * as styles from 'styles';
+import {makeStyles} from "@material-ui/core";
+
+const useStyles = makeStyles(theme => ({
+  root: {
+    margin: '1vh',
+    backgroundColor: '#5b6bc0',
+    width: '20vh'
+  }
+}));
 
 const CopyButton = props => {
   const [isComplete, setIsComplete] = useState(false);
   const [message, setMessage] = useState('Copy to clipboard');
+  const classes = useStyles();
 
   const saveToClipboard = () => {
     let element = document.createElement("input");
@@ -18,10 +27,10 @@ const CopyButton = props => {
   };
 
   return (
-    <Button url={props.url}
+    <Button className={classes.root}
+            url={props.url}
             variant="contained"
             color="primary"
-            style={styles.app.button}
             disabled={document.queryCommandEnabled('copy')}
             onClick={!isComplete ? saveToClipboard : undefined}
             href={null}>
